% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cmaGetters.R
\name{cmaSetDimension}
\alias{cmaGetCountEval}
\alias{cmaGetDimension}
\alias{cmaGetInitialX}
\alias{cmaGetPopulationSize}
\alias{cmaSetCountEval}
\alias{cmaSetDimension}
\alias{cmaSetInitialX}
\alias{cmaSetPopulationSize}
\title{rCMA Getters and Setters.}
\usage{
cmaSetDimension(cma, i)

cmaGetDimension(cma)

cmaSetPopulationSize(cma, i)

cmaGetPopulationSize(cma)

cmaSetInitialX(cma, initialX)

cmaGetInitialX(cma)

cmaSetCountEval(cma, p)

cmaGetCountEval(cma)
}
\arguments{
\item{cma}{CMA-ES Java object, created with \code{\link{cmaNew}}}

\item{i}{a parameter of type integer}

\item{initialX}{either a double or a double vector of length  \code{\link{cmaGetDimension}}}

\item{p}{a parameter of type long}
}
\value{
none for the setters, the requested element(s) for the getters
}
\description{
Get or set various elements of CMA-ES Java object \code{cma}. \cr
\cr
\code{cmaSetDimension} sets the problem dimension (only prior to \code{\link{cmaInit}}) \cr
\code{cmaGetDimension} returns the problem dimension \cr
\code{cmaSetPopulationSize} sets the population size (only prior to \code{\link{cmaInit}}) \cr
\code{cmaGetPopulationSize} returns the population size  \cr
\code{cmaSetInitialX} set the mean vector for the initial population (only prior to \code{\link{cmaInit}}) \cr
\code{cmaGetInitialX} returns the mean vector for the initial population  \cr
\code{cmaSetCountEval} sets the counter for fitness function evaluations (only prior to \code{\link{cmaInit}}) \cr
\code{cmaGetCountEval} returns the counter for fitness function evaluations  \cr
}
\seealso{
\code{\link{cmaSetStopFitness}}, \code{\link{cmaNew}}, \code{\link{cmaInit}}
}

