% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rCMA.R
\name{cmaUpdateDistribution}
\alias{cmaUpdateDistribution}
\title{Update CMA-ES distribution with the fitness vector of the last population.}
\usage{
cmaUpdateDistribution(cma, fitness)
}
\arguments{
\item{cma}{CMA-ES Java object, already initialized with \code{\link{cmaInit}}}

\item{fitness}{vector of length \code{\link{cmaGetPopulationSize}(cma)} with the fitness of each individuum}
}
\description{
Update CMA-ES distribution with the fitness vector of the last population.
}
\note{
As a side effect, the CMA-ES Java object cma of class \code{CMAEvolutionStrategy}
 is augmented.
}
\author{
Wolfgang Konen, FHK, 2013
}
\seealso{
\code{\link{cmaSamplePopulation}}, \code{\link{cmaNew}}, \code{\link{cmaOptimDP}}
}

